campus_lbs/requirements.txt
Django==5.0.6
psycopg[binary]==3.2.1
djangorestframework==3.15.2
djangorestframework-gis==1.1
django-filter==24.2
django-environ==0.11.2
gunicorn==22.0.0
geos==0.2.3
